refactor(das): remove das-types dependency and integrate types directly into rpc
