+++ rpc/Cargo.toml
@@ -19,7 +19,6 @@ categories = [
 ]
 
 [dependencies]
-das-types = { path = "./types" }
 celestia-types.workspace = true
 prost.workspace = true
