+++ rpc/tests/das.rs
@@ -1,6 +1,7 @@
 #![cfg(not(target_arch = "wasm32"))]
 
 use celestia_rpc::prelude::*;
+use celestia_rpc::das::SamplingStats;
 
 pub mod utils;
 
@@ -11,14 +12,14 @@ async fn das_sampling_stats() {
     let client = new_test_client(AuthLevel::Read).await.unwrap();
 
     let stats1 = client.das_sampling_stats().await.unwrap();
+    assert!(matches!(stats1, SamplingStats { .. }));
 
     client
         .header_wait_for_height(stats1.network_head_height + 1)
         .await
         .unwrap();
 
-    let stats2 = client.das_sampling_stats().await.unwrap();
-
+    let stats2: SamplingStats = client.das_sampling_stats().await.unwrap();
     assert!(stats2.head_of_sampled_chain >= stats1.head_of_sampled_chain);
 }
