--- rpc/types/Cargo.toml
@@ -1,47 +1,0 @@
-[package]
-name = "das-types"
-version = "0.10.1"
-edition = "2021"
-license = "Apache-2.0"
-description = "Core types for Data Availability Sampling (DAS) in the Celestia ecosystem"
-authors = ["Eiger <hello@eiger.co>"]
-homepage = "https://www.eiger.co"
-repository = "https://github.com/eigerco/lumina"
-readme = "README.md"
-keywords = ["blockchain", "celestia", "das"]
-categories = ["encoding", "cryptography::cryptocurrencies"]
-
-[dependencies]
-libp2p-identity = { version = "0.2.9", optional = true }
-multiaddr = { version = "0.18.1", optional = true }
-serde = { version = "1.0.203", features = ["derive"] }
-serde_repr = "0.1.19"
-sha2 = "0.10.6"
-thiserror = "1.0.61"
-time = { version = "0.3.36", default-features = false }
-
-# wasm-bindgen related dependencies (optional)
-js-sys = { version = "0.3.76", optional = true }
-serde-wasm-bindgen = { version = "0.6.5", optional = true }
-wasm-bindgen = { workspace = true, optional = true }
-nmt-rs = { workspace = true, features = ["serde"], optional = true }
-
-[dev-dependencies]
-serde_json = "1.0.117"
-bincode = "1.3.3"
-indoc = "2.0.5"
-
-[target.'cfg(target_arch = "wasm32")'.dev-dependencies]
-wasm-bindgen-test.workspace = true
-
-[features]
-default = ["p2p"]
-p2p = ["dep:libp2p-identity", "dep:multiaddr"]
-wasm-bindgen = ["dep:js-sys", "dep:serde-wasm-bindgen", "dep:wasm-bindgen", "dep:nmt-rs", "time/wasm-bindgen"]
-
-[package.metadata.docs.rs]
-features = ["p2p", "test-utils"]
-rustdoc-args = ["--cfg", "docsrs"]
-
-[package.metadata.cargo-udeps.ignore]
-development = ["indoc"]
