--- rpc/types/src/lib.rs
@@ -1,6 +1,0 @@
-#![cfg_attr(docsrs, feature(doc_cfg))]
-#![doc = include_str!("../../README.md")]
-
-mod das;
-
-pub use crate::das::*;
